LayoutTests/editing/pasteboard/smart-paste-006.html

<!DOCTYPE html>
<html> 
<head>
<style>
.editing { 
    border: 2px solid red; 
    font-size: 24px; 
}
.explanation { 
    border: 2px solid blue; 
    padding: 12px; 
    font-size: 24px; 
    margin-bottom: 24px;
}
.scenario { margin-bottom: 16px;}
.scenario:first-line { font-weight: bold; margin-bottom: 16px;}
.expected-results:first-line { font-weight: bold }
</style>
</head> 
<body>

<div class="explanation">
<div class="scenario">
Tests: 
<br>
Smart paste when pasting between smart paste exempt characters.
</div>
<div class="expected-results">
Expected Results:
<br>
No spaces should be added before or after the pasted word. It should look like this:
<BR>
-test-
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="test" class="editing">
test
</div>
</div>
<script src="../editing.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script>

async function editingTest() {
    await UIHelper.selectWordByDoubleTapOrClick(document.getElementById('test'));    cutCommand();
    typeCharacterCommand('-');
    typeCharacterCommand('-');
    moveSelectionBackwardByCharacterCommand();
    pasteCommand();
}

runDumpAsTextEditingTest(true);
</script>
</body>
</html>
